feat(tryout): prepare canonical response contraction
