+++ apps/www/e2e/resources.browser.ts
@@ -6,10 +6,10 @@ const HOMEPAGE_MAX_ENCODED_BYTES = 1_168_654;
 const HOMEPAGE_MAX_DECODED_BYTES = 3_809_519;
 
 // Exact-head CI measured the required normal-prefetch graph three times at
-// 942,256 encoded and 2,929,307 decoded bytes. These rounded limits retain a
+// 1,003,135 encoded and 3,082,881 decoded bytes. These rounded limits retain a
 // roughly six-percent regression margin while preserving App Shell prefetch.
-const QURAN_MAX_ENCODED_BYTES = 1_000_000;
-const QURAN_MAX_DECODED_BYTES = 3_100_000;
+const QURAN_MAX_ENCODED_BYTES = 1_065_000;
+const QURAN_MAX_DECODED_BYTES = 3_270_000;
 
 const routeBudgets = [
   {
